treewide: format all inactive Nix files
Inactive means that there are no open PRs with activity in the
last month that touch those files.
A bunch later, we can do another pass to get the rest.

Doing it this way makes ensures that we don't cause any conflicts for
recently active PRs that would be ready to merge.
Only once those PRs get updated, CI will kick in and require the files
to be formatted.

Furthermore, this makes sure that we can merge this PR without having
to constantly rebase it!

This can be verified using

    nix-build https://gist.github.com/infinisil/4b7a1a1e5db681d04446e73e39048aec/archive/5439e16eda66856aed747db3b45ed595835de7c6.tar.gz \
      --argstr baseRev 50a5bef72eecf24c0f176ed6d3ce856077e1f61a
    result/bin/apply-formatting $NIXPKGS_PATH

+++ pkgs/applications/networking/p2p/deluge/default.nix
@@ -1,13 +1,14 @@
-{ lib
-, fetchurl
-, intltool
-, libtorrent-rasterbar
-, python3Packages
-, gtk3
-, glib
-, gobject-introspection
-, librsvg
-, wrapGAppsHook3
+{
+  lib,
+  fetchurl,
+  intltool,
+  libtorrent-rasterbar,
+  python3Packages,
+  gtk3,
+  glib,
+  gobject-introspection,
+  librsvg,
+  wrapGAppsHook3,
 }:
 
 let
@@ -15,7 +16,8 @@ let
 
   pypkgs = python3Packages;
 
-  generic = { pname, withGUI }:
+  generic =
+    { pname, withGUI }:
     pypkgs.buildPythonPackage rec {
       inherit pname;
       version = "2.1.1";
@@ -25,37 +27,42 @@ let
         hash = "sha256-do3TGYAuQkN6s3lOvnW0lxQuCO1bD7JQO61izvRC3/c=";
       };
 
-      propagatedBuildInputs = with pypkgs; [
-        twisted
-        mako
-        chardet
-        pyxdg
-        pyopenssl
-        service-identity
-        libtorrent-rasterbar.dev
-        libtorrent-rasterbar.python
-        setuptools
-        setproctitle
-        pillow
-        rencode
-        six
-        zope-interface
-        dbus-python
-        pycairo
-        librsvg
-      ] ++ optionals withGUI [
-        gtk3
-        gobject-introspection
-        pygobject3
-      ];
+      propagatedBuildInputs =
+        with pypkgs;
+        [
+          twisted
+          mako
+          chardet
+          pyxdg
+          pyopenssl
+          service-identity
+          libtorrent-rasterbar.dev
+          libtorrent-rasterbar.python
+          setuptools
+          setproctitle
+          pillow
+          rencode
+          six
+          zope-interface
+          dbus-python
+          pycairo
+          librsvg
+        ]
+        ++ optionals withGUI [
+          gtk3
+          gobject-introspection
+          pygobject3
+        ];
 
-      nativeBuildInputs = [
-        intltool
-        glib
-      ] ++ optionals withGUI [
-        gobject-introspection
-        wrapGAppsHook3
-      ];
+      nativeBuildInputs =
+        [
+          intltool
+          glib
+        ]
+        ++ optionals withGUI [
+          gobject-introspection
+          wrapGAppsHook3
+        ];
 
       nativeCheckInputs = with pypkgs; [
         pytestCheckHook
@@ -68,18 +75,24 @@ let
 
       doCheck = false; # tests are not working at all
 
-      postInstall = ''
-        install -Dm444 -t $out/lib/systemd/system packaging/systemd/*.service
-      '' + (if withGUI
-      then ''
-        mkdir -p $out/share
-        cp -R deluge/ui/data/{icons,pixmaps} $out/share/
-        install -Dm444 -t $out/share/applications deluge/ui/data/share/applications/deluge.desktop
-      '' else ''
-        rm -r $out/bin/deluge-gtk
-        rm -r $out/${python3Packages.python.sitePackages}/deluge/ui/gtk3
-        rm -r $out/share/{icons,man/man1/deluge-gtk*,pixmaps}
-      '');
+      postInstall =
+        ''
+          install -Dm444 -t $out/lib/systemd/system packaging/systemd/*.service
+        ''
+        + (
+          if withGUI then
+            ''
+              mkdir -p $out/share
+              cp -R deluge/ui/data/{icons,pixmaps} $out/share/
+              install -Dm444 -t $out/share/applications deluge/ui/data/share/applications/deluge.desktop
+            ''
+          else
+            ''
+              rm -r $out/bin/deluge-gtk
+              rm -r $out/${python3Packages.python.sitePackages}/deluge/ui/gtk3
+              rm -r $out/share/{icons,man/man1/deluge-gtk*,pixmaps}
+            ''
+        );
 
       postFixup = ''
         for f in $out/lib/systemd/system/*; do
@@ -91,14 +104,23 @@ let
         description = "Torrent client";
         homepage = "https://deluge-torrent.org";
         license = licenses.gpl3Plus;
-        maintainers = with maintainers; [ domenkozar ebzzry ];
+        maintainers = with maintainers; [
+          domenkozar
+          ebzzry
+        ];
         platforms = platforms.all;
       };
     };
 
 in
 rec {
-  deluge-gtk = generic { pname = "deluge-gtk"; withGUI = true; };
-  deluged = generic { pname = "deluged"; withGUI = false; };
+  deluge-gtk = generic {
+    pname = "deluge-gtk";
+    withGUI = true;
+  };
+  deluged = generic {
+    pname = "deluged";
+    withGUI = false;
+  };
   deluge = deluge-gtk;
 }
