+++ pkgs/os-specific/linux/atop/default.nix
@@ -1,13 +1,14 @@
-{ lib
-, stdenv
-, fetchurl
-, zlib
-, ncurses
-, findutils
-, systemd
-, python3
-# makes the package unfree via pynvml
-, withAtopgpu ? false
+{
+  lib,
+  stdenv,
+  fetchurl,
+  zlib,
+  ncurses,
+  findutils,
+  systemd,
+  python3,
+  # makes the package unfree via pynvml
+  withAtopgpu ? false,
 }:
 
 stdenv.mkDerivation rec {
@@ -19,20 +20,14 @@ stdenv.mkDerivation rec {
     sha256 = "sha256-lwBYoZt5w0RPlx+FRXKg5jiR3C1fcDf/g3VwhUzg2h4=";
   };
 
-  nativeBuildInputs = lib.optionals withAtopgpu [
-    python3.pkgs.wrapPython
-  ];
+  nativeBuildInputs = lib.optionals withAtopgpu [ python3.pkgs.wrapPython ];
 
   buildInputs = [
     zlib
     ncurses
-  ] ++ lib.optionals withAtopgpu [
-    python3
-  ];
+  ] ++ lib.optionals withAtopgpu [ python3 ];
 
-  pythonPath = lib.optionals withAtopgpu [
-    python3.pkgs.pynvml
-  ];
+  pythonPath = lib.optionals withAtopgpu [ python3.pkgs.pynvml ];
 
   makeFlags = [
     "DESTDIR=$(out)"
@@ -68,14 +63,21 @@ stdenv.mkDerivation rec {
     mkdir -p $out/bin
   '';
 
-  postInstall = ''
-    # Remove extra files we don't need
-    rm -r $out/{var,etc} $out/bin/atop{sar,}-${version}
-  '' + (if withAtopgpu then ''
-    wrapPythonPrograms
-  '' else ''
-    rm $out/lib/systemd/system/atopgpu.service $out/bin/atopgpud $out/share/man/man8/atopgpud.8
-  '');
+  postInstall =
+    ''
+      # Remove extra files we don't need
+      rm -r $out/{var,etc} $out/bin/atop{sar,}-${version}
+    ''
+    + (
+      if withAtopgpu then
+        ''
+          wrapPythonPrograms
+        ''
+      else
+        ''
+          rm $out/lib/systemd/system/atopgpu.service $out/bin/atopgpud $out/share/man/man8/atopgpud.8
+        ''
+    );
 
   meta = with lib; {
     platforms = platforms.linux;
